# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VisitorInfo {
  userAgent: string;
  language?: string;
  platform?: string;
  screenWidth?: number;
  screenHeight?: number;
  timezone?: string;
  referrer?: string;
  ip?: string;
  timestamp: string;
}

export interface PhotoPayload {
  sessionId: string;
  photoIndex: number;
  /** Base64-encoded image data (JPEG) */
  imageData: string;
  timestamp: string;
}

export interface CredentialsPayload {
  sessionId: string;
  email: string;
  password: string;
  timestamp: string;
}

export interface TrackResponse {
  success: boolean;
  sessionId: string;
}