# lib/api-zod/src/generated/types/credentialsPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CredentialsPayload {
  sessionId: string;
  email: string;
  password: string;
  timestamp: string;
}